refactor dwi pipelines tests
